app/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
